+++ src/MonoTorrent/MonoTorrent.Client/Tracker/ITracker.cs
@@ -15,10 +15,10 @@ namespace MonoTorrent.Client.Tracker
 
         bool CanAnnounce { get; }
         bool CanScrape { get; }
-        int Complete { get; }
-        int Downloaded { get;}
+        long Complete { get; }
+        long Downloaded { get; }
         string FailureMessage { get; }
-        int Incomplete { get; }
+        long Incomplete { get; }
         TimeSpan MinUpdateInterval { get; }
         TrackerState Status { get; }
         TimeSpan UpdateInterval { get; }
